refactor(service/endpoints): remove routers module and consolidate endpoints to reduce complexity and improve maintainability (#74)
* fix(service/schemas): update InputData class to correct typo in units field

* refactor(service/endpoints): consolidate router endpoints into a single module to reduce complexity and centralize endpoints

* refactor(service/main): update code format to comply with flake8 quality check
