+++ service/src/schemas.py
@@ -7,7 +7,7 @@ class InputData(BaseModel):
     sex: Literal["male", "female"] = Field(
         description="The sex of the individual ('male' or 'female')."
     )
-    units: Literal["si", "imaperial"] = Field(
+    units: Literal["si", "imperial"] = Field(
         description="The units system ('si' or 'imperial')."
     )
     age: int = Field(
